Refactoring. Optimizing dependency graph.
`apply` of `HasMethodMatching` was only used by `Typers` and `unapply` of `HasMethodMatching` was only used by `Implicits`. Allthough they represent the same type of thing and can be used together they are actually not the same thing. The name `HasMethodMatching` does not match the contents of the `apply` method.

Another option would be to rename this combination to something that represent the construction. I do not think that this construction deserves it's own type as it is only used in a very limited number of places. On top of that, the unapply matches more types than are created by the apply method.

- Moved `apply` method to a new private object `MemberWildcardMethodType` in `Typers`
- Made supporting constructs `dummyMethod` and `templateArgType` private
- Made `HasMethodMatching` private